Add more functionality to mainWindow

+++ app/src/main/java/com/example/weatherforecast/api/WeatherResponse.kt
@@ -7,7 +7,7 @@ data class WeatherResponse(
 )
 
 data class Main(
-    val temp: Double,
+    val temp: Float,
     val pressure: Int,
     val humidity: Int
 )
